requirements.txt
Pillow==10.3.0
opencv-python==4.9.0.80
pyzbar==0.1.9
pytest==8.3.1
# tkinter and sqlite3 are part of the Python standard installation on Raspberry Pi OS
